python-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
opencv-python>=4.10.0.84
Pillow>=10.4.0
numpy>=1.26.4
mediapipe>=0.10.14